Moved AddNote concenrn functions into it
